+++ src/constants.ts
@@ -1,3 +1,1 @@
 export const EXTENSION_NAME = "gitdoc";
-export const ENABLED_KEY = `${EXTENSION_NAME}:enabled`;
-export const BRANCH_ENABLED_KEY = `${EXTENSION_NAME}:branchEnabled`;