+++ CHANGELOG.md
@@ -1,6 +1,13 @@
-## v0.0.4 (04/18/2020)
+## v0.0.5 (04/23/2020)
 
-- Introduced the `Collapse Version(s) Above` command to the `Timeline` tree, which allows you to merge/name a bunch of auto-commits into a single version
+- Introduced the `Undo Version` command to the `Timeline` tree, which allows you to revert a previous change
+- Added the new `GitDoc: Auto Commit Delay` setting, which allows you to control how much time to delay creating commits after saving
+- Renamed the `Collapse Version(s) Above` command to `Squash Version(s) Above`
+- Renamed the `onSave` option of the `GitDoc: Auto Push` setting to `onCommit`
+
+## v0.0.4 (04/19/2020)
+
+- Introduced the `Squash Version(s) Above` command to the `Timeline` tree, which allows you to merge/name a bunch of auto-commits into a single version
 - Introduced the `Restore Version` command to the `Timeline` tree, which allows you to restore a previous version of a file
 
 ## v0.0.3 (04/18/2020)